Update lock
Use conditional expression with shape

Correct typo in test

Update RealField to work with arrays

Add the inverse Gaussian distribution

Bump

Split structure classes into 3 files

Polish parameter_supports mecahnism

Factor out triangular number functions

Skip sampling test for inverse Gaussian

Correct issue with scalar support bound shapes

Split parameter package

Correct docstrings

Bump

Eliminate Jax initialization on import

Update examples

Bump

change sample using NP directly

resolve conflicts

Remove commented-out code related to Inverse Gaussian in pytest_generate_tests

resolve conflicts

resolve conflicts
