+++ .pre-commit-config.yaml
@@ -15,11 +15,11 @@ repos:
   - id: check-yaml
 
 - repo: https://github.com/astral-sh/ruff-pre-commit
-  rev: v0.11.3
+  rev: v0.11.5
   hooks:
   - id: ruff
 
 - repo: https://github.com/RobertCraigie/pyright-python
-  rev: v1.1.398
+  rev: v1.1.399
   hooks:
   - id: pyright
