+++ efax/_src/distributions/cmvn/unit_variance.py
@@ -118,10 +118,7 @@ class ComplexMultivariateUnitVarianceNormalEP(
 
     @override
     def sample(self, key: KeyArray, shape: Shape | None = None) -> JaxComplexArray:
-        if shape is not None:
-            shape += self.mean.shape
-        else:
-            shape = self.mean.shape
+        shape = self.mean.shape if shape is None else shape + self.mean.shape
         a = jr.normal(key, shape)
         b = jr.normal(key, shape)
         return a + 1j * b + self.mean
