+++ efax/_src/distributions/geometric.py
@@ -79,9 +79,9 @@ class GeometricEP(HasEntropyEP[GeometricNP],
 
     @override
     def sample(self, key: KeyArray, shape: Shape | None = None) -> JaxRealArray:
-        xp = self.array_namespace()
         if shape is not None:
             shape += self.shape
+        xp = self.array_namespace()
         p = xp.reciprocal(self.mean)
         return jr.geometric(key, p, shape)
