+++ efax/_src/distributions/log_normal/unit_variance.py
@@ -68,10 +68,7 @@ class UnitVarianceLogNormalNP(
     @override
     def sample(self, key: KeyArray, shape: Shape | None = None) -> JaxRealArray:
         xp = self.array_namespace()
-        if shape is not None:
-            shape += self.shape
-        else:
-            shape = self.shape
+        shape = self.shape if shape is None else shape + self.shape
         return jr.lognormal(key, shape=shape) * xp.exp(self.mean)
 
 
@@ -118,10 +115,7 @@ class UnitVarianceLogNormalEP(
     @override
     def sample(self, key: KeyArray, shape: Shape | None = None) -> JaxRealArray:
         xp = self.array_namespace()
-        if shape is not None:
-            shape += self.shape
-        else:
-            shape = self.shape
+        shape = self.shape if shape is None else shape + self.shape
         return jr.lognormal(key, shape=shape) * xp.exp(self.mean)
 
     @override
