+++ efax/_src/distributions/multivariate_normal/arbitrary.py
@@ -168,10 +168,7 @@ class MultivariateNormalVP(Samplable, Multidimensional):
 
     @override
     def sample(self, key: KeyArray, shape: Shape | None = None) -> JaxRealArray:
-        if shape is not None:
-            shape += self.shape
-        else:
-            shape = self.shape
+        shape = self.shape if shape is None else shape + self.shape
         return jr.multivariate_normal(key, self.mean, self.variance, shape)
 
     @override
