+++ efax/_src/distributions/multivariate_normal/fixed_variance.py
@@ -126,11 +126,8 @@ class MultivariateFixedVarianceNormalEP(
 
     @override
     def sample(self, key: KeyArray, shape: Shape | None = None) -> JaxRealArray:
+        shape = self.mean.shape if shape is None else shape + self.mean.shape
         xp = self.array_namespace()
-        if shape is not None:
-            shape += self.mean.shape
-        else:
-            shape = self.mean.shape
         variance = self.variance[..., xp.newaxis]
         deviation = xp.sqrt(variance)
         return jr.normal(key, shape) * deviation + self.mean
