+++ efax/_src/distributions/multivariate_normal/isotropic.py
@@ -119,11 +119,8 @@ class IsotropicNormalEP(HasEntropyEP[IsotropicNormalNP],
 
     @override
     def sample(self, key: KeyArray, shape: Shape | None = None) -> JaxRealArray:
+        shape = self.mean.shape if shape is None else shape + self.mean.shape
         xp = self.array_namespace()
-        if shape is not None:
-            shape += self.mean.shape
-        else:
-            shape = self.mean.shape
         deviation = xp.sqrt(self.variance())
         return jr.normal(key, shape) * deviation[..., xp.newaxis] + self.mean
