+++ efax/_src/distributions/normal/unit_variance.py
@@ -63,10 +63,7 @@ class UnitVarianceNormalNP(HasEntropyNP['UnitVarianceNormalEP'],
 
     @override
     def sample(self, key: KeyArray, shape: Shape | None = None) -> JaxRealArray:
-        if shape is not None:
-            shape += self.mean.shape
-        else:
-            shape = self.mean.shape
+        shape = self.shape if shape is None else shape + self.shape
         return jr.normal(key, shape) + self.mean
