+++ efax/_src/distributions/poisson.py
@@ -100,10 +100,7 @@ class PoissonEP(HasConjugatePrior,
 
     @override
     def sample(self, key: KeyArray, shape: Shape | None = None) -> JaxRealArray:
-        if shape is not None:
-            shape += self.shape
-        else:
-            shape = self.shape
+        shape = self.shape if shape is None else shape + self.shape
         return jr.poisson(key, self.mean, shape)
 
     @override
