+++ efax/_src/distributions/softplus_normal/unit_variance.py
@@ -63,10 +63,7 @@ class UnitVarianceSoftplusNormalNP(
 
     @override
     def sample(self, key: KeyArray, shape: Shape | None = None) -> JaxRealArray:
-        if shape is not None:
-            shape += self.shape
-        else:
-            shape = self.shape
+        shape = self.shape if shape is None else shape + self.shape
         return softplus(jr.normal(key, shape) + self.mean)
 
 
@@ -106,10 +103,7 @@ class UnitVarianceSoftplusNormalEP(
 
     @override
     def sample(self, key: KeyArray, shape: Shape | None = None) -> JaxRealArray:
-        if shape is not None:
-            shape += self.shape
-        else:
-            shape = self.shape
+        shape = self.shape if shape is None else shape + self.shape
         return softplus(jr.normal(key, shape) + self.mean)
 
     @override
