+++ efax/_src/tools.py
@@ -13,7 +13,7 @@ from tjax import JaxComplexArray, JaxRealArray
 
 from .iteration import parameters
 from .parametrization import Distribution
-from .structure import Structure
+from .structure.structure import Structure
 from .types import Axis
 
 
@@ -41,7 +41,7 @@ def parameter_mean(x: T, /, *, axis: Axis | None = None) -> T:
     """Return the mean of the parameters (fixed and variable)."""
     xp = x.array_namespace()
     structure = Structure.create(x)
-    p = parameters(x, support=False)
+    p = parameters(x)
     q = {path: xp.mean(value, axis=axis) for path, value in p.items()}
     return structure.assemble(q)
