+++ examples/.editorconfig
@@ -1,0 +1,2 @@
+[*.py]
+max_line_length = 80
