+++ examples/optimization.py
@@ -1,9 +1,9 @@
 """Optimization.
 
-This example illustrates how this library fits in a typical machine learning context.  Suppose we
-have an unknown target value, and a loss function based on the cross-entropy between the target
-value and a predictive distribution.  We will optimize the predictive distribution by a small
-fraction of its cotangent.
+This example illustrates how this library fits in a typical machine learning
+context.  Suppose we have an unknown target value, and a loss function based on
+the cross-entropy between the target value and a predictive distribution.  We
+will optimize the predictive distribution by a small fraction of its cotangent.
 """
 import jax.numpy as jnp
 from jax import grad, lax
@@ -34,30 +34,36 @@ def cond_fun(q: BernoulliNP) -> JaxBooleanArray:
     return total > 1e-6  # noqa: PLR2004
 
 
-# The target_distribution is represented as the expectation parameters of a Bernoulli distribution
-# corresponding to probabilities 0.3, 0.4, and 0.7.
+# The target_distribution is represented as the expectation parameters of a
+# Bernoulli distribution corresponding to probabilities 0.3, 0.4, and 0.7.
 target_distribution = BernoulliEP(jnp.asarray([0.3, 0.4, 0.7]))
 
-# The initial predictive distribution is represented as the natural parameters of a Bernoulli
-# distribution corresponding to log-odds 0, which is probability 0.5.
+# The initial predictive distribution is represented as the natural parameters
+# of a Bernoulli distribution corresponding to log-odds 0, which is probability
+# 0.5.
 initial_predictive_distribution = BernoulliNP(jnp.zeros(3))
 
-# Optimize the predictive distribution iteratively, and output the natural parameters of the
-# prediction.
-predictive_distribution = lax.while_loop(cond_fun, body_fun, initial_predictive_distribution)
-print_generic(predictive_distribution)
-# BernoulliNP
+# Optimize the predictive distribution iteratively.
+predictive_distribution = lax.while_loop(cond_fun, body_fun,
+                                         initial_predictive_distribution)
+
+# Compare the optimized predictive distribution with the target value in the
+# same natural parametrization.
+print_generic(predictive_distribution=predictive_distribution,
+              target_distribution=target_distribution.to_nat())
+# predictive_distribution=BernoulliNP[dataclass]
 # └── log_odds=Jax Array (3,) float32
 #     └──  -0.8440 │ -0.4047 │ 0.8440
-
-# Compare the optimized predictive distribution with the target value in the same parametrization.
-print_generic(target_distribution.to_nat())
-# BernoulliNP
+# target_distribution=BernoulliNP[dataclass]
 # └── log_odds=Jax Array (3,) float32
 #     └──  -0.8473 │ -0.4055 │ 0.8473
 
-# Print the optimized natural parameters as expectation parameters.
-print_generic(predictive_distribution.to_exp())
-# BernoulliEP
+# Do the same in the expectation parametrization.
+print_generic(predictive_distribution=predictive_distribution.to_exp(),
+              target_distribution=target_distribution)
+# predictive_distribution=BernoulliEP[dataclass]
 # └── probability=Jax Array (3,) float32
 #     └──  0.3007 │ 0.4002 │ 0.6993
+# target_distribution=BernoulliEP[dataclass]
+# └── probability=Jax Array (3,) float32
+#     └──  0.3000 │ 0.4000 │ 0.7000
